+++ CHANGES.md
@@ -15,7 +15,7 @@
     addressed by KnuckleDragger's induction methods. If you were using this functionality and have problems
     porting to knuckle-dragger, please get in touch!
 
-   * Add function 'last' to Data.SBV.List.
+   * Add functions 'last' and 'replicate' to Data.SBV.List.
 
    * Add Documentation.SBV.Examples.KnuckleDragger.Reverse, proving a curious definition of reverse that
      uses no auxiliary definitions is correct.